requirements.txt
insightface==0.6.2
numpy==1.22.3
opencv_python_headless==4.5.5.64
pandas==1.4.2
Pillow==9.1.0
pydantic==1.9.0
scikit_learn==1.0.2
SQLAlchemy==1.4.35
streamlit==1.8.1
